lab05 q3 test

+++ lab05/snapshot-load.sh
@@ -3,6 +3,8 @@
 
 counter_bp=$1;
 
+# notice this place should use the ../
+# This is due to the error print by the terminal in vlab
 ../snapshot-save.sh
 
 echo "Restoring snapshot $counter_bp"
@@ -15,8 +17,19 @@ do
         # avoiding copy the same current file
         if [ "$file" != "snapshot-load.sh" ]
 	    then
-            
-			rm $file
+            if [ "$file" != "backup.sh" ]
+	        then
+                if [ "$file" != "backup.pl" ]
+	            then
+			        rm $file
+                else
+                    # pass
+                    :
+                fi
+            else
+                # pass
+                :
+            fi
         else
             # pass
             :
